This is a part of large PR https://github.com/dmlc/xgboost/pull/7208. Here only the changes in ColumnMatrix are implemented.
